Working on ColissionDetection

+++ input.py
@@ -18,5 +18,5 @@ def checkForInput(mario):
     if pygame.mouse.get_pressed()[0]:
         mario.vel.x = 0
         mario.vel.y = 0       
-        mario.pos.x = pygame.mouse.get_pos()[0]/32
-        mario.pos.y = pygame.mouse.get_pos()[1]/32
+        mario.pos.x = int(pygame.mouse.get_pos()[0]/32)
+        mario.pos.y = int(pygame.mouse.get_pos()[1]/32)
